Fixed issue #9932
If an ext_pillar item is not loaded and its branch
is not present in pillar_roots, the corresponding
salt environment will not be available. However,
this is not an issue, since it will be available
when all of ext_pillar items are parsed; therefore,
when trying to get the SLS from that non-present
salt environment, we should merely log a debug
message and continue on.